+++ library/src/schemas/object/object.ts
@@ -84,7 +84,7 @@ export function object<TObjectShape extends ObjectShape>(
      */
     _parse(input, info) {
       // Check type of input
-      if (!input || input.constructor !== Object) {
+      if (!input || typeof input !== 'object') {
         return getIssues(
           info,
           'type',