+++ library/src/schemas/object/objectAsync.ts
@@ -96,7 +96,7 @@ export function objectAsync<TObjectShape extends ObjectShapeAsync>(
      */
     async _parse(input, info) {
       // Check type of input
-      if (!input || input.constructor !== Object) {
+      if (!input || typeof input !== 'object') {
         return getIssues(
           info,
           'type',