+++ library/src/schemas/record/recordAsync.ts
@@ -141,7 +141,7 @@ export function recordAsync<
      */
     async _parse(input, info) {
       // Check type of input
-      if (!input || input.constructor !== Object) {
+      if (!input || typeof input !== 'object') {
         return getIssues(
           info,
           'type',